Killing a non-existent query is harmless, so don't die.
There is a race condition between the moment the processlist is received
and the moment an order to kill a process id is issued. The query in
question may have finished in the meantime, in which case it's no longer
there to be killed. If this happens, then just ignore it and carry on.